refactor : 페이지 네임 변경
