+++ frontend/src/App.js
@@ -1,22 +1,22 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import './App.css';
-import LoginPage from "./pages/LoginPage";
-import LandingPage from "./pages/LandingPage";
-import CreateVoice from "./pages/CreateVoice";
-import VoiceMarket from "./pages/VoiceMarket";
-import SignupPage from "./pages/SignupPage"
+import SignIn from "./pages/SignIn";
+import SignUp from "./pages/SignUp"
+import Landing from "./pages/Landing";
+import VoiceCreate from "./pages/VoiceCreate";
+import VoiceStore from "./pages/VoiceStore";
 
 function App() {
   return (
       <Router>
         <Routes>
-            <Route path="/login" element={<LoginPage />} />
-            <Route path="/signup" element={<SignupPage/>}/>
-            <Route path="/landingpage" element={<LandingPage />} />
-            <Route path="/createvoice" element={<CreateVoice />} />
-            <Route path="/voicemarket" element={<VoiceMarket />} />
+            <Route path="/signin" element={<SignIn />} />
+            <Route path="/signup" element={<SignUp/>}/>
+            <Route path="/landing" element={<Landing />} />
+            <Route path="/voicecreate" element={<VoiceCreate />} />
+            <Route path="/voicestore" element={<VoiceStore />} />
             {/* 기본 경로를 로그인 페이지로 설정 */}
-            <Route path="*" element={<LoginPage />} />
+            <Route path="*" element={<SignIn />} />
         </Routes>
       </Router>
   );
