+++ frontend/src/pages/Landing.js
@@ -22,7 +22,7 @@ const LandingPage = () => {
   }, []);
 
   const handleCreateVoice = () => {
-    navigate('/createVoice');
+    navigate('/voicecreate');
   };
 
   return (
@@ -35,7 +35,7 @@ const LandingPage = () => {
           보이스팩 생성
         </button>
         <button 
-          onClick={() => navigate('/voiceMarket')}
+          onClick={() => navigate('/voicestore')}
           className="bg-purple-500 hover:bg-purple-700 text-white font-bold py-2 px-4 rounded-md w-80"
         >
           보이스팩 구매
